\name{Example Datasets}
\alias{pstats}
\alias{pval.2d.complex}
\alias{pval.2d.mag}
\alias{shepp0fMRI}
\alias{shepp1fMRI}
\alias{shepp2fMRI}
\alias{sheppAnat}
\alias{toy1}
\alias{toy2}
\docType{data}
\title{Example datasets in MixfMRI}
\description{
  These are datasets used to demo examples and workflows in this package.
}
\format{
  Objects may contain several information and data.
}
\details{
  \code{pstats} is a 3D example.

  \code{pval.2d.complex} and \code{pval.2d.mag} are 2D examples.

  \code{shepp0fMRI}, \code{shepp1fMRI}, \code{shepp2fMRI} and
  \code{sheppAnat} are phantoms generated by Dr. Maitra
  for simulation studies with different overlap levels for p-values.

  \code{toy1} and \code{toy2} are two 3D toy examples.
}
\references{
  Chen, W.-C. and Maitra, R. (2021)
  \dQuote{A Practical Model-based Segmentation Approach for Accurate
	  Activation Detection in Single-Subject functional Magnetic
	  Resonance Imaging Studies},
  \emph{arXiv:2102.03639}.
}
\author{
  Wei-Chen Chen and Ranjan Maitra.
}
\examples{
library(MixfMRI, quietly = TRUE)

### Plotting.
demo(shepp,'MixfMRI',ask=FALSE,echo=FALSE)
}
\keyword{datasets}
