\name{General Internal Functions}
\alias{.MixfMRIEnv}
\alias{my.YlOrRd}
\alias{my.YlGnBu}
\alias{my.Reds}
\alias{my.alpha.append}
\alias{print.lmt.I}
\alias{print.lmt.pv}
\title{All General Internal Functions and Datasets}
\description{
  All general internal functions
}
\references{
  \url{http://maitra.public.iastate.edu/}
}
\author{
  Wei-Chen Chen and Ranjan Maitra.
}
\keyword{internal}
