\name{General Internal Functions}
\alias{.MixfMRIEnv}
\alias{my.YlOrRd}
\alias{my.YlGnBu}
\alias{my.Reds}
\alias{my.alpha.append}
\alias{print.lmt.I}
\alias{print.lmt.pv}
\title{All General Internal Functions and Datasets}
\description{
  All general internal functions
}
\references{
  Chen, W.-C. and Maitra, R. (2021)
  \dQuote{A Practical Model-based Segmentation Approach for Accurate
	  Activation Detection in Single-Subject functional Magnetic
	  Resonance Imaging Studies},
  \emph{arXiv:2102.03639}.
}
\author{
  Wei-Chen Chen and Ranjan Maitra.
}
\keyword{internal}
